chore: rename interactive legend caption sub components
