--- src/components/instant-apps-interactive-legend/instant-apps-interactive-legend-caption/instant-apps-interactive-legend-caption.scss
@@ -1,71 +1,0 @@
-:host {
-  display: block;
-  --instant-apps-interactive-legend-heading-font-size: 1.25rem;
-  --instant-apps-interactive-legend-heading-font-weight: 800;
-  --instant-apps-interactive-legend-secondary-color: var(--calcite-ui-text-1);
-  --instant-apps-interactive-legend-ui-padding: 15px 10px;
-
-  .esri-widget__heading {
-    overflow: hidden;
-  }
-
-  .esri-widget__heading {
-    margin: 0;
-    overflow-x: hidden;
-  }
-
-  .esri-widget__heading {
-    display: inline-block;
-    font-size: var(--instant-apps-interactive-legend-heading-font-size);
-  }
-
-  .esri-widget__heading {
-    font-weight: var(--instant-apps-interactive-legend-heading-font-weight);
-  }
-
-  .esri-widget__heading header {
-    padding: var(--instant-apps-interactive-legend-ui-padding);
-  }
-
-  header {
-    display: flex;
-    text-align: left;
-    padding: var(--instant-apps-interactive-legend-ui-padding);
-
-    h3 {
-      margin-bottom: 0;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      display: -webkit-box;
-      -webkit-line-clamp: 2; /* number of lines to show */
-      line-clamp: 2;
-      -webkit-box-orient: vertical;
-      line-height: normal;
-    }
-
-    .instant-apps-interactive-legend__heading-text--group-item {
-      -webkit-line-clamp: 1; /* number of lines to show */
-      line-clamp: 1;
-    }
-
-    span {
-      width: 100%;
-      font-size: var(--instant-apps-interactive-legend-total-feature-count-font-size);
-      color: var(--instant-apps-interactive-legend-secondary-color);
-    }
-
-    .instant-apps-interactive-legend__header-action-container {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-  }
-
-  .esri-widget__heading header {
-    padding: var(--instant-apps-interactive-legend-ui-padding);
-  }
-
-  .esri-widget--panel {
-    margin: 0;
-  }
-}
