+++ src/components/instant-apps-interactive-legend/instant-apps-interactive-legend-legend-element-caption/instant-apps-interactive-legend-legend-element-caption.scss
@@ -1,0 +1,70 @@
+:host {
+  display: block;
+
+  --instant-apps-interactive-legend-heading-font-size: 1.25rem;
+  --instant-apps-interactive-legend-heading-font-weight: 800;
+  --instant-apps-interactive-legend-caption-font-weight: 450;
+  --instant-apps-interactive-legend-secondary-color: var(--calcite-ui-text-1);
+  --instant-apps-interactive-legend-field-name-font-size: 1rem;
+  --instant-apps-interactive-legend-total-feature-count-font-size: 0.8125rem;
+  --instant-apps-interactive-legend-info-font-size: 0.875rem;
+  --instant-apps-interactive-legend-info-item-background--selected: #c7ebff;
+  --instant-apps-interactive-legend-info-item-color--selected: var(--calcite-ui-text-2);
+  --instant-apps-interactive-legend-secondary-background-color: var(--calcite-ui-background);
+  --instant-apps-interactive-legend-ui-padding: 15px 10px;
+
+  .esri-legend__layer-caption {
+    calcite-action {
+      margin-left: 5px;
+    }
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: var(--instant-apps-interactive-legend-field-name-font-size);
+    background-color: var(--instant-apps-interactive-legend-secondary-background-color);
+    color: var(--instant-apps-interactive-legend-secondary-color);
+
+    calcite-action {
+      margin-left: auto;
+    }
+  }
+
+  .instant-apps-interactive-legend__legend-layer-caption-text {
+    display: inline-block;
+    margin-left: 5px;
+    margin-bottom: 0;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 1; /* number of lines to show */
+    line-clamp: 1;
+    -webkit-box-orient: vertical;
+    line-height: normal;
+  }
+
+  .instant-apps-interactive-legend__layer-caption-btn-container {
+    white-space: nowrap;
+    calcite-button {
+      margin: 2px;
+    }
+  }
+
+  .esri-legend__layer-caption {
+    font-weight: var(--instant-apps-interactive-legend-caption-font-weight);
+  }
+
+  .esri-legend__layer-caption {
+    padding: var(--instant-apps-interactive-legend-ui-padding);
+  }
+}
+
+html[dir='rtl'] {
+  :host {
+    .esri-legend__layer-caption {
+      calcite-action {
+        margin-left: unset;
+        margin-right: auto;
+      }
+    }
+  }
+}
