+++ src/components/instant-apps-interactive-legend/instant-apps-interactive-legend-legend-element-caption/readme.md
@@ -1,0 +1,58 @@
+# instant-apps-interactive-legend-legend-element-caption
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Properties
+
+| Property             | Attribute              | Description | Type              | Default     |
+| -------------------- | ---------------------- | ----------- | ----------------- | ----------- |
+| `activeLayerInfo`    | --                     |             | `ActiveLayerInfo` | `undefined` |
+| `expanded`           | `expanded`             |             | `boolean`         | `true`      |
+| `isInteractive`      | `is-interactive`       |             | `boolean`         | `undefined` |
+| `layer`              | --                     |             | `FeatureLayer`    | `undefined` |
+| `legendElementIndex` | `legend-element-index` |             | `number`          | `undefined` |
+| `legendvm`           | --                     |             | `LegendViewModel` | `undefined` |
+| `messages`           | `messages`             |             | `any`             | `undefined` |
+| `titleText`          | `title-text`           |             | `string`          | `undefined` |
+| `zoomTo`             | `zoom-to`              |             | `boolean`         | `undefined` |
+
+
+## Events
+
+| Event                      | Description | Type                   |
+| -------------------------- | ----------- | ---------------------- |
+| `legendLayerExpandUpdated` |             | `CustomEvent<boolean>` |
+| `showAllSelected`          |             | `CustomEvent<boolean>` |
+
+
+## Dependencies
+
+### Used by
+
+ - [instant-apps-interactive-legend-legend-element](../instant-apps-interactive-legend-legend-element)
+
+### Graph
+```mermaid
+graph TD;
+  instant-apps-interactive-legend-legend-element --> instant-apps-interactive-legend-legend-element-caption
+  style instant-apps-interactive-legend-legend-element-caption fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+## License
+COPYRIGHT © 2023 Esri
+
+All rights reserved under the copyright laws of the United States and applicable international laws, treaties, and conventions.
+
+This material is licensed for use under the Esri Master License Agreement (MLA), and is bound by the terms of that agreement. You may redistribute and use this code without modification, provided you adhere to the terms of the MLA and include this copyright notice.
+
+See use restrictions at http://www.esri.com/legal/pdfs/mla_e204_e300/english
+
+For additional information, contact: Environmental Systems Research Institute, Inc. Attn: Contracts and Legal Services Department 380 New York Street Redlands, California, USA 92373 USA
+
+email: contracts@esri.com
+
