+++ src/components/instant-apps-interactive-legend/instant-apps-interactive-legend-legend-element/readme.md
@@ -1,0 +1,58 @@
+# instant-apps-interactive-legend-legend-element
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Properties
+
+| Property             | Attribute              | Description | Type                                                                                                                               | Default     |
+| -------------------- | ---------------------- | ----------- | ---------------------------------------------------------------------------------------------------------------------------------- | ----------- |
+| `activeLayerInfo`    | --                     |             | `ActiveLayerInfo`                                                                                                                  | `undefined` |
+| `isChild`            | `is-child`             |             | `boolean`                                                                                                                          | `undefined` |
+| `isColorRamp`        | `is-color-ramp`        |             | `boolean`                                                                                                                          | `undefined` |
+| `isInteractive`      | `is-interactive`       |             | `boolean`                                                                                                                          | `undefined` |
+| `isRelationshipRamp` | `is-relationship-ramp` |             | `boolean`                                                                                                                          | `undefined` |
+| `isSizeRamp`         | `is-size-ramp`         |             | `boolean`                                                                                                                          | `undefined` |
+| `legendElement`      | --                     |             | `ColorRampElement \| HeatmapRampElement \| OpacityRampElement \| RelationshipRampElement \| SizeRampElement \| SymbolTableElement` | `undefined` |
+| `legendElementIndex` | `legend-element-index` |             | `number`                                                                                                                           | `undefined` |
+| `legendvm`           | --                     |             | `LegendViewModel`                                                                                                                  | `undefined` |
+| `messages`           | `messages`             |             | `any`                                                                                                                              | `undefined` |
+| `titleText`          | `title-text`           |             | `string`                                                                                                                           | `undefined` |
+| `zoomTo`             | `zoom-to`              |             | `boolean`                                                                                                                          | `undefined` |
+
+
+## Dependencies
+
+### Used by
+
+ - [instant-apps-interactive-legend-classic](../instant-apps-interactive-legend-classic)
+
+### Depends on
+
+- [instant-apps-interactive-legend-legend-element-caption](../instant-apps-interactive-legend-legend-element-caption)
+
+### Graph
+```mermaid
+graph TD;
+  instant-apps-interactive-legend-legend-element --> instant-apps-interactive-legend-legend-element-caption
+  instant-apps-interactive-legend-classic --> instant-apps-interactive-legend-legend-element
+  style instant-apps-interactive-legend-legend-element fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+## License
+COPYRIGHT © 2023 Esri
+
+All rights reserved under the copyright laws of the United States and applicable international laws, treaties, and conventions.
+
+This material is licensed for use under the Esri Master License Agreement (MLA), and is bound by the terms of that agreement. You may redistribute and use this code without modification, provided you adhere to the terms of the MLA and include this copyright notice.
+
+See use restrictions at http://www.esri.com/legal/pdfs/mla_e204_e300/english
+
+For additional information, contact: Environmental Systems Research Institute, Inc. Attn: Contracts and Legal Services Department 380 New York Street Redlands, California, USA 92373 USA
+
+email: contracts@esri.com
+
